merge develop to release

+++ src/services/auth.service.ts
@@ -2,51 +2,39 @@ import httpStatus from 'http-status';
 import tokenService from './token.service';
 import userService from './user.service';
 import ApiError from '../utils/ApiError';
-import { TokenType, User } from '@prisma/client';
+import { TokenType } from '@prisma/client';
 import prisma from '../client';
 import { encryptPassword, isPasswordMatch } from '../utils/encryption';
 import { AuthTokensResponse } from '../types/response';
 import exclude from '../utils/exclude';
+import { IUser } from '../interfaces/user.interface';
 
-/**
- * Login with username and password
- * @param {string} email
- * @param {string} password
- * @returns {Promise<Omit<User, 'password'>>}
- */
 const loginUserWithEmailAndPassword = async (
   email: string,
   password: string
-): Promise<
-  Omit<
-    Pick<
-      User,
-      'id' | 'email' | 'name' | 'password' | 'role' | 'isEmailVerified' | 'createdAt' | 'updatedAt'
-    >,
-    'password'
-  >
-> => {
+): Promise<Omit<IUser, 'password'>> => {
   const user = await userService.getUserByEmail(email, [
     'id',
     'email',
     'name',
+    'avatar',
     'password',
     'role',
     'isEmailVerified',
     'createdAt',
-    'updatedAt'
+    'updatedAt',
+    'height',
+    'weight',
+    'age'
   ]);
+
   if (!user || !(await isPasswordMatch(password, user.password as string))) {
-    throw new ApiError(httpStatus.UNAUTHORIZED, 'Incorrect email or password');
+    throw new ApiError(httpStatus.UNPROCESSABLE_ENTITY, 'Email hoặc mật khẩu không chính xác');
   }
-  return exclude(user, ['password']);
+
+  return exclude(user, ['password']) as Omit<IUser, 'password'>;
 };
 
-/**
- * Logout
- * @param {string} refreshToken
- * @returns {Promise<void>}
- */
 const logout = async (refreshToken: string): Promise<void> => {
   const refreshTokenData = await prisma.token.findFirst({
     where: {
@@ -55,69 +43,66 @@ const logout = async (refreshToken: string): Promise<void> => {
       blacklisted: false
     }
   });
+
   if (!refreshTokenData) {
-    throw new ApiError(httpStatus.NOT_FOUND, 'Not found');
+    throw new ApiError(httpStatus.NOT_FOUND, 'Không tìm thấy refresh token');
   }
+
   await prisma.token.delete({ where: { id: refreshTokenData.id } });
 };
 
-/**
- * Refresh auth tokens
- * @param {string} refreshToken
- * @returns {Promise<AuthTokensResponse>}
- */
 const refreshAuth = async (refreshToken: string): Promise<AuthTokensResponse> => {
   try {
     const refreshTokenData = await tokenService.verifyToken(refreshToken, TokenType.REFRESH);
-    const { userId } = refreshTokenData;
+    const user = await userService.getUserById(refreshTokenData.userId, ['id', 'isEmailVerified']);
+
+    if (!user) {
+      throw new Error();
+    }
+
     await prisma.token.delete({ where: { id: refreshTokenData.id } });
-    return tokenService.generateAuthTokens({ id: userId });
+    return tokenService.generateAuthTokens({
+      id: user.id,
+      isEmailVerified: user.isEmailVerified
+    });
   } catch (error) {
-    throw new ApiError(httpStatus.UNAUTHORIZED, 'Please authenticate');
+    throw new ApiError(httpStatus.UNAUTHORIZED, 'Vui lòng đăng nhập');
   }
 };
 
-/**
- * Reset password
- * @param {string} resetPasswordToken
- * @param {string} newPassword
- * @returns {Promise<void>}
- */
 const resetPassword = async (resetPasswordToken: string, newPassword: string): Promise<void> => {
   try {
     const resetPasswordTokenData = await tokenService.verifyToken(
       resetPasswordToken,
       TokenType.RESET_PASSWORD
     );
+
     const user = await userService.getUserById(resetPasswordTokenData.userId);
     if (!user) {
       throw new Error();
     }
-    const encryptedPassword = await encryptPassword(newPassword);
-    await userService.updateUserById(user.id, { password: encryptedPassword });
+
+    await userService.updateUserById(user.id, { password: newPassword });
     await prisma.token.deleteMany({ where: { userId: user.id, type: TokenType.RESET_PASSWORD } });
   } catch (error) {
-    throw new ApiError(httpStatus.UNAUTHORIZED, 'Password reset failed');
+    throw new ApiError(httpStatus.BAD_REQUEST, 'Đặt lại mật khẩu thất bại hoặc token không hợp lệ');
   }
 };
 
-/**
- * Verify email
- * @param {string} verifyEmailToken
- * @returns {Promise<void>}
- */
 const verifyEmail = async (verifyEmailToken: string): Promise<void> => {
   try {
     const verifyEmailTokenData = await tokenService.verifyToken(
       verifyEmailToken,
       TokenType.VERIFY_EMAIL
     );
+
     await prisma.token.deleteMany({
       where: { userId: verifyEmailTokenData.userId, type: TokenType.VERIFY_EMAIL }
     });
+
     await userService.updateUserById(verifyEmailTokenData.userId, { isEmailVerified: true });
   } catch (error) {
-    throw new ApiError(httpStatus.UNAUTHORIZED, 'Email verification failed');
+    throw new ApiError(httpStatus.BAD_REQUEST, 'Xác thực email thất bại hoặc token không hợp lệ');
   }
 };
